Rename type.h -> cu_file.h
