+++ src/supervisor.cpp
@@ -13,13 +13,13 @@
 #include "so_loading.h"
 #include "symbolication/callstack.h"
 #include "symbolication/cu.h"
+#include "symbolication/cu_file.h"
 #include "symbolication/dwarf_expressions.h"
 #include "symbolication/dwarf_frameunwinder.h"
 #include "symbolication/elf.h"
 #include "symbolication/elf_symbols.h"
 #include "symbolication/lnp.h"
 #include "symbolication/objfile.h"
-#include "symbolication/type.h"
 #include "task.h"
 #include "tracer.h"
 #include "utils/logger.h"
