+++ src/supervisor.h
@@ -7,8 +7,8 @@
 #include "ptracestop_handlers.h"
 #include "so_loading.h"
 #include "symbolication/callstack.h"
+#include "symbolication/cu_file.h"
 #include "symbolication/elf.h"
-#include "symbolication/type.h"
 #include "task.h"
 #include "utils/static_vector.h"
 #include <algorithm>
