+++ src/symbolication/callstack.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "../common.h"
-#include "type.h"
+#include "cu_file.h"
 #include <vector>
 
 namespace ui::dap {
