+++ src/symbolication/cu.h
@@ -1,9 +1,9 @@
 #pragma once
 
+#include "cu_file.h"
 #include "dwarf.h"
 #include "lnp.h"
 #include "objfile.h"
-#include "type.h"
 #include <concepts>
 #include <cstddef>
 #include <cstring>
