+++ src/symbolication/cu_file.cpp
@@ -1,4 +1,4 @@
-#include "type.h"
+#include "cu_file.h"
 #include "block.h"
 #include "dwarf.h"
 #include "elf.h"
