+++ src/symbolication/disassemble.cpp
@@ -1,10 +1,10 @@
 #include "disassemble.h"
 #include "../supervisor.h"
+#include "cu_file.h"
 #include "elf.h"
 #include "fmt/core.h"
 #include "lnp.h"
 #include "objfile.h"
-#include "type.h"
 #include "zydis/Zydis.h"
 #include <algorithm>
 #include <charconv>
